backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
grpcio==1.66.0
grpcio-tools==1.66.0
httpx==0.27.0
openai==1.50.0
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-json-logger==3.1.0
